Try to mitigate some timing related test problems.

+++ tests/Delay.hs
@@ -13,15 +13,15 @@ timedDelay msg usecs = do
 
 thr1 :: IO ()
 thr1 = do
-  timedDelay "thr1" 100000
+  timedDelay "thr1" (100000::Int)
 
 thr2 :: IO ()
 thr2 = do
-  timedDelay "thr2" 300000
+  timedDelay "thr2" (300000::Int)
 
 main :: IO ()
 main = do
   forkIO thr1
   forkIO thr2
-  timedDelay "main" 200000
-  timedDelay "end" 500000
+  timedDelay "main" (200000::Int)
+  timedDelay "end" (500000::Int)
